+++ lib/modules/state/state_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: state
-// Updated at: 2026-06-06 09:13:54
-// Commit: feat(state): integrate offline caching for state state
+// Updated at: 2026-06-06 20:30:57
+// Commit: feat(state): implement responsive state view component
 
 class stateComponentState {
-  final String commitRef = '1800';
-  final String lastMessage = 'feat(state): integrate offline caching for state state';
-  final DateTime timestamp = DateTime.parse('2026-06-06T09:13:54+05:00');
+  final String commitRef = '1817';
+  final String lastMessage = 'feat(state): implement responsive state view component';
+  final DateTime timestamp = DateTime.parse('2026-06-06T20:30:57+05:00');
 }